refactor(record): consolidate record APIs by resource
