--- apps/web/src/entities/record/api/emotion-post.ts
@@ -1,14 +1,0 @@
-import { api } from '@/shared/lib/ky'
-import type { EmotionResponseDTO } from '@/entities/record/model/emotion.type'
-
-export const emotionPost = {
-  postEmotionRecord: async (
-    matchRecordId: number,
-    emotionType: string,
-  ): Promise<EmotionResponseDTO> => {
-    const response = await api
-      .post(`emotion`, { json: { matchRecordId, emotionType } })
-      .json<EmotionResponseDTO>()
-    return response
-  },
-}
