+++ apps/web/src/entities/record/api/emotion.queries.ts
@@ -1,6 +1,6 @@
 import { createQueryKeys } from '@lukemorales/query-key-factory'
 
-import { emotionGet } from './emotion-get'
+import { emotionGet } from './emotion.api'
 
 export const emotions = createQueryKeys('emotion', {
   record: (matchRecordId: number) => ({
