+++ apps/web/src/entities/record/api/record.api.ts
@@ -1,9 +1,10 @@
 import { api } from '@/shared/lib/ky'
 
 import type {
-  RecordResponseDTO,
+  RecordDeleteResponseDTO,
   RecordDetailResponseDTO,
   RecordEmotionStatsResponseDTO,
+  RecordResponseDTO,
 } from '../model/record.type'
 
 export const recordGet = {
@@ -25,6 +26,15 @@ export const recordGet = {
     const response = await api
       .get(`record/matches/${matchId}/team`)
       .json<RecordEmotionStatsResponseDTO>()
+    return response
+  },
+}
+
+export const recordDelete = {
+  deleteRecord: async (recordId: number): Promise<RecordDeleteResponseDTO> => {
+    const response = await api
+      .delete(`record/${recordId}`)
+      .json<RecordDeleteResponseDTO>()
     return response
   },
 }
