--- apps/web/src/entities/record/api/recording-get.ts
@@ -1,12 +1,0 @@
-import { api } from '@/shared/lib/ky'
-import type { RecordingResponseDTO } from '@/entities/record/model/recording.type'
-
-export const recordingGet = {
-  getRecording: async (matchId: number): Promise<RecordingResponseDTO> => {
-    const response = await api
-      .get(`record/${matchId}/match`, { retry: 0 })
-      .json<RecordingResponseDTO>()
-
-    return response
-  },
-}
