+++ apps/web/src/entities/record/index.ts
@@ -1,0 +1,4 @@
+export * from './api'
+export * from './model'
+export * from './ui'
+export * from './constants/record'
