+++ apps/web/src/entities/record/model/index.ts
@@ -1,0 +1,25 @@
+export type {
+  MatchResult,
+  RecordResult,
+  RecordPatchResult,
+  Record,
+  RecordResponse,
+  Image,
+  EmotionGroup,
+  RecordDetailResponse,
+  RecordEmotionStatsResponse,
+  RecordResponseDTO,
+  RecordDetailResponseDTO,
+  RecordDeleteResponseDTO,
+  RecordEmotionStatsResponseDTO,
+} from './record.type'
+
+export type {
+  RecordingResponse,
+  RecordingPostResponse,
+  RecordingResponseDTO,
+  RecordingPatchResponseDTO,
+  RecordingPostResponseDTO,
+} from './recording.type'
+
+export type { EmotionType, EmotionResponseDTO } from './emotion.type'
